ElementsWidget: Move ElementFilterList to implementation
========================================================

Only moving code from header to implementation.
